Improved Slate to Markdown example
With the current example the passed in `onChange` callback would be called on mount as well as every time the prop changes, which is probably not what a user wants. (compare for example to <input>).
Instead we should call it only when the editor's `onChange` is called, i.e. when the state is supposed to update.