fix: updated NodeTypes to fix customization error.
The node types were set in a way that caused customization to be challenging. For example, the ol_list was typed as the string literal "ol_list", when using typescript this caused issues with types not matching. Since the defaultNodeTypes is used it will have the default types all set. Setting it as a string will allow others to customize the values.